Allow server edition from dashboard
When there is a connection problem before the server has been
successfully set up, it is convenient to be able to edit it directly
from the dashboard.

Also added a dashboard link to go to the list of your servers if you
have no active servers but do have inactive ones.