Re-run setup when digests have changed
The digest of the playbook is now computed based on all the files in the
playbook directory. The digest of the variables is also computed and
checked, so that a change in class configuration (e.g. teacher ssh
public keys to install on servers) also triggers a re-run of the setup
playbook.