Avoid infinite crash/ansible loop
If the server manager crashes while running the ansible setup playbook,
it might get stuck in a loop where it repeatedly crashes then re-runs
the playbook. If the last three setup playbook runs were not successful,
a problem is added to the state and the playbook is not automatically
run again. It can still be run manually.