Fix class form servers enabled flag toggle
Either there's a bug in how the form is implemented or there's a bug in
live view. The current implementation somehow toggles the servers
enabled flag to true when there's any other change in the form. On first
analysis, this seems to be because the value comes back as "true" or
"false" (a binary) in the form field, which always evaluates as truthy.
Must investigate and file a bug report if appropriate.