Show cloud server data in course
This commit adds a card containing the cloud server setup data (or
details once it's been created) in the course, either in the page on
small screens or under the table of contents on large screens.

The card contains the basic information necessary to set up the server
in the virtual server setup exercise, or the server's connection details
for further exercises.