Stagger server connections over a few seconds
When many servers servers try to connect at the same time, e.g. when the
application starts or when a class is enabled, the applications has
trouble keeping up with the sudden load.

These changes introduces a 2-7 seconds delay to the server manager on
initial connection and reactivation, staggering reconnections over a
period of 5 seconds to avoid too much load at the same time.